Move GeneraleMissiven.to_documents() to new abstract Sheet class.
